Question 1.5. SW1 and SW2 functionality
Used functions display_on_LEDs and display_on_LCD to display the binary value of count of the LEDs and the decimal value of count on the LCD. SW1 increments the count value by 1 (updating the display and LEDs by 1) and SW2 decrements the count value by (updating the display and LEDs by 1). On Reset, the value of 0 is displayed on the LEDs and the LCD. The code does not exceed the limits of count i.e. when count of 0 is reached and decrement this value, 255 is displayed and when 255 is displayed and increment count, 0 is displayed. The limits to count have not been exceed.